requirements.txt
fastapi
python-telegram-bot
supabase
python-dotenv
uvicorn